code review for validators
